Move all egress to dotnet-monitor.
