+++ diagnostics.sln
@@ -175,8 +175,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Diagnostics.Monit
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Diagnostics.Monitoring.EventPipe.UnitTests", "src\tests\Microsoft.Diagnostics.Monitoring.EventPipe\Microsoft.Diagnostics.Monitoring.EventPipe.UnitTests.csproj", "{A58B7F72-E9DF-4962-9E15-300EF4C53ABB}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Diagnostics.Monitoring.Egress.AzureStorage", "src\Microsoft.Diagnostics.Monitoring.Egress.AzureStorage\Microsoft.Diagnostics.Monitoring.Egress.AzureStorage.csproj", "{519BBD81-A555-46F1-A331-233FD21123D7}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Checked|Any CPU = Checked|Any CPU
@@ -1410,46 +1408,6 @@ Global
 		{A58B7F72-E9DF-4962-9E15-300EF4C53ABB}.RelWithDebInfo|x64.Build.0 = Release|Any CPU
 		{A58B7F72-E9DF-4962-9E15-300EF4C53ABB}.RelWithDebInfo|x86.ActiveCfg = Release|Any CPU
 		{A58B7F72-E9DF-4962-9E15-300EF4C53ABB}.RelWithDebInfo|x86.Build.0 = Release|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.Checked|Any CPU.ActiveCfg = Debug|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.Checked|Any CPU.Build.0 = Debug|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.Checked|ARM.ActiveCfg = Debug|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.Checked|ARM.Build.0 = Debug|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.Checked|ARM64.ActiveCfg = Debug|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.Checked|ARM64.Build.0 = Debug|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.Checked|x64.ActiveCfg = Debug|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.Checked|x64.Build.0 = Debug|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.Checked|x86.ActiveCfg = Debug|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.Checked|x86.Build.0 = Debug|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.Debug|ARM.ActiveCfg = Debug|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.Debug|ARM.Build.0 = Debug|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.Debug|ARM64.ActiveCfg = Debug|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.Debug|ARM64.Build.0 = Debug|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.Debug|x64.Build.0 = Debug|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.Debug|x86.Build.0 = Debug|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.Release|Any CPU.Build.0 = Release|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.Release|ARM.ActiveCfg = Release|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.Release|ARM.Build.0 = Release|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.Release|ARM64.ActiveCfg = Release|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.Release|ARM64.Build.0 = Release|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.Release|x64.ActiveCfg = Release|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.Release|x64.Build.0 = Release|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.Release|x86.ActiveCfg = Release|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.Release|x86.Build.0 = Release|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.RelWithDebInfo|Any CPU.ActiveCfg = Release|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.RelWithDebInfo|Any CPU.Build.0 = Release|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.RelWithDebInfo|ARM.ActiveCfg = Release|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.RelWithDebInfo|ARM.Build.0 = Release|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.RelWithDebInfo|ARM64.ActiveCfg = Release|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.RelWithDebInfo|ARM64.Build.0 = Release|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.RelWithDebInfo|x64.ActiveCfg = Release|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.RelWithDebInfo|x64.Build.0 = Release|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.RelWithDebInfo|x86.ActiveCfg = Release|Any CPU
-		{519BBD81-A555-46F1-A331-233FD21123D7}.RelWithDebInfo|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -1495,7 +1453,6 @@ Global
 		{870A7AD0-B808-491E-92F6-433BE990B374} = {03479E19-3F18-49A6-910A-F5041E27E7C0}
 		{E1E2C698-E247-4D48-AE01-542B43744ACB} = {19FAB78C-3351-4911-8F0C-8C6056401740}
 		{A58B7F72-E9DF-4962-9E15-300EF4C53ABB} = {03479E19-3F18-49A6-910A-F5041E27E7C0}
-		{519BBD81-A555-46F1-A331-233FD21123D7} = {19FAB78C-3351-4911-8F0C-8C6056401740}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {46465737-C938-44FC-BE1A-4CE139EBB5E0}
