--- src/Microsoft.Diagnostics.Monitoring.Egress.AzureStorage/Microsoft.Diagnostics.Monitoring.Egress.AzureStorage.csproj
@@ -1,30 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFrameworks>netstandard2.0</TargetFrameworks>
-    <NoWarn>;1591;1701</NoWarn>
-    <Description>Egress for dotnet-monitor</Description>
-    <!-- Tentatively create package so other teams can tentatively consume. -->
-    <IsPackable>true</IsPackable>
-    <PackageTags>Diagnostic</PackageTags>
-    <PackageReleaseNotes>$(Description)</PackageReleaseNotes>
-    <GenerateDocumentationFile>false</GenerateDocumentationFile>
-    <IncludeSymbols>true</IncludeSymbols>
-    <!-- Do not ship this package until ready to be consumed. -->
-    <IsShipping>false</IsShipping>
-    <OutputType>Library</OutputType>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Azure.Storage.Blobs" Version="$(AzureStorageBlobsVersion)" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <InternalsVisibleTo Include="dotnet-monitor" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Microsoft.Diagnostics.Monitoring\Microsoft.Diagnostics.Monitoring.csproj" />
-  </ItemGroup>
-
-</Project>
