+++ src/Tools/dotnet-monitor/Egress/Configuration/EgressConfigurationExtensions.cs
@@ -4,7 +4,7 @@
 
 using Microsoft.Extensions.Configuration;
 
-namespace Microsoft.Diagnostics.Tools.Monitor
+namespace Microsoft.Diagnostics.Tools.Monitor.Egress.Configuration
 {
     internal static class EgressConfigurationExtensions
     {
