+++ src/Tools/dotnet-monitor/Egress/Configuration/EgressConfigureOptions.cs
@@ -9,7 +9,7 @@ using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 
-namespace Microsoft.Diagnostics.Tools.Monitor
+namespace Microsoft.Diagnostics.Tools.Monitor.Egress.Configuration
 {
     /* 
      * == Egress Configuration Design ==
