+++ src/Tools/dotnet-monitor/Egress/Configuration/EgressOptions.cs
@@ -5,7 +5,7 @@
 using System;
 using System.Collections.Generic;
 
-namespace Microsoft.Diagnostics.Tools.Monitor
+namespace Microsoft.Diagnostics.Tools.Monitor.Egress.Configuration
 {
     /// <summary>
     /// Configuration options for specifying egress providers.
