+++ src/Tools/dotnet-monitor/Egress/ConfiguredEgressProvider.cs
@@ -9,7 +9,7 @@ using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Microsoft.Diagnostics.Tools.Monitor
+namespace Microsoft.Diagnostics.Tools.Monitor.Egress
 {
     /// <summary>
     /// Base class for configured egress providers. A configured egress provider is an egress provider instance
