+++ src/Tools/dotnet-monitor/Egress/FileSystem/FileSystemEgressStreamOptions.cs
@@ -2,7 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace Microsoft.Diagnostics.Monitoring.Egress.FileSystem
+namespace Microsoft.Diagnostics.Tools.Monitor.Egress.FileSystem
 {
     /// <summary>
     /// Egress stream options for file system egress.
